Added state transitions and reworked some of the state code.
The enemy crab can now transition between its idle and pursuit state. Also the code that stores the different states of the enemy have been rewritten. Instead of there being a hardcoded instance of each state, there is now a dictionary which stores all of the possible AI states for the enemy.